+++ pkg/agent/ebpfspy/ebpfspy_linux.go
@@ -24,8 +24,8 @@ func NewEBPFSpy(s *Session) *EbpfSpy {
 	}
 }
 
-func Start(pid int, _ spy.ProfileType, sampleRate uint32, _ bool) (spy.Spy, error) {
-	s := NewSession(pid, sampleRate, 256, sd.NoopServiceDiscovery{}, false)
+func Start(params spy.InitParams) (spy.Spy, error) {
+	s := NewSession(params.Logger, params.Pid, params.SampleRate, 256, sd.NoopServiceDiscovery{}, false)
 	err := s.Start()
 	if err != nil {
 		return nil, err