+++ pkg/agent/rbspy/rbspy.go
@@ -36,7 +36,7 @@ type RbSpy struct {
 	pid int
 }
 
-func Start(pid int, _ spy.ProfileType, _ uint32, _ bool) (spy.Spy, error) {
+func Start(params spy.InitParams) (spy.Spy, error) {
 	dataBuf := make([]byte, bufferLength)
 	dataPtr := unsafe.Pointer(&dataBuf[0])
 
@@ -51,7 +51,7 @@ func Start(pid int, _ spy.ProfileType, _ uint32, _ bool) (spy.Spy, error) {
 	if Blocking {
 		blocking = 1
 	}
-	r := C.rbspy_init(C.int(pid), C.int(blocking), errorPtr, C.int(bufferLength))
+	r := C.rbspy_init(C.int(params.Pid), C.int(blocking), errorPtr, C.int(bufferLength))
 
 	if r < 0 {
 		return nil, errors.New(string(errorBuf[:-r]))
@@ -62,7 +62,7 @@ func Start(pid int, _ spy.ProfileType, _ uint32, _ bool) (spy.Spy, error) {
 		dataBuf:  dataBuf,
 		errorBuf: errorBuf,
 		errorPtr: errorPtr,
-		pid:      pid,
+		pid:      params.Pid,
 	}, nil
 }