+++ pkg/storage/storage_put.go
@@ -26,9 +26,6 @@ type PutInput struct {
 }
 
 func (s *Storage) Put(ctx context.Context, pi *PutInput) error {
-	// TODO: This is a pretty broad lock. We should find a way to make these locks more selective.
-	s.putMutex.Lock()
-	defer s.putMutex.Unlock()
 	if s.hc.IsOutOfDiskSpace() {
 		return errOutOfSpace
 	}
@@ -53,8 +50,8 @@ func (s *Storage) Put(ctx context.Context, pi *PutInput) error {
 		"aggregationType": pi.AggregationType,
 	}).Debug("storage.Put")
 
-	for k, v := range pi.Key.Labels() {
-		s.labels.Put(k, v)
+	if err := s.labels.PutLabels(pi.Key.Labels()); err != nil {
+		return fmt.Errorf("unable to write labels: %w", err)
 	}
 
 	sk := pi.Key.SegmentKey()
@@ -75,9 +72,14 @@ func (s *Storage) Put(ctx context.Context, pi *PutInput) error {
 	}
 
 	st := r.(*segment.Segment)
-	st.SetMetadata(pi.SpyName, pi.SampleRate, pi.Units, pi.AggregationType)
-	samples := pi.Val.Samples()
+	st.SetMetadata(metadata.Metadata{
+		SpyName:         pi.SpyName,
+		SampleRate:      pi.SampleRate,
+		Units:           pi.Units,
+		AggregationType: pi.AggregationType,
+	})
 
+	samples := pi.Val.Samples()
 	err = st.Put(pi.StartTime, pi.EndTime, samples, func(depth int, t time.Time, r *big.Rat, addons []segment.Addon) {
 		tk := pi.Key.TreeKey(depth, t)
 		res, err := s.trees.GetOrCreate(tk)