+++ packages/pyroscope-panel-plugin/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.56](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.55...@pyroscope/panel-plugin@1.3.56) (2022-08-30)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.3.55](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.54...@pyroscope/panel-plugin@1.3.55) (2022-08-30)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.3.54](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.53...@pyroscope/panel-plugin@1.3.54) (2022-08-26)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
 ## [1.3.53](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.52...@pyroscope/panel-plugin@1.3.53) (2022-08-22)
 
 **Note:** Version bump only for package @pyroscope/panel-plugin